refactor : 상수를 의미에 맞는 예외클래스로 옮겨라
