+++ src/main/java/com/codessquad/qna/web/domain/Question.java
@@ -8,7 +8,9 @@ import org.hibernate.annotations.Where;
 import javax.persistence.*;
 import java.util.List;
 
-import static com.codessquad.qna.web.utils.ExceptionConstants.*;
+import static com.codessquad.qna.web.exceptions.InvalidEntityException.EMPTY_FIELD_IN_QUESTION_ENTITY;
+import static com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException.CANNOT_MODIFY_OR_DELETE_ANOTHER_USERS_QUESTION;
+import static com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException.CAN_NOT_DELETE_BECAUSE_ANOTHER_USERS_ANSWER_IS_EXISTS;
 
 @Entity
 public class Question extends BaseTimeEntity {
